+++ src/pages/collect/index.tsx
@@ -5,8 +5,8 @@ import { BackButtonNavigation } from "@/common/components/Navigation";
 import { MemeListSkeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
 import { useDebounce, useInput } from "@/common/hooks";
-import { withAuth } from "@/components/hocs";
 import { Collection, SearchedCollection } from "@/features/collect/components";
+import { withAuth } from "@/features/common";
 import { SearchInput } from "@/features/search/components";
 
 const CollectPage = () => {
