+++ src/pages/share/index.tsx
@@ -4,7 +4,7 @@ import { useAuth } from "@/application/hooks";
 import { BackButtonNavigation } from "@/common/components/Navigation";
 import { MemeListSkeleton } from "@/common/components/Skeleton";
 import { SSRSuspense } from "@/common/components/Suspense";
-import { withAuth } from "@/components/hocs";
+import { withAuth } from "@/features/common";
 import { SharedMemeList } from "@/features/share/components";
 
 const SharedHistoryPage: NextPage = () => {
